backend/requirements.txt
# ─────────────────────────────────────────────────────────
# RAG Pipeline — Python Dependencies
# Python 3.11.1
# ─────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9          # multipart/form-data file uploads

# ── Database ──────────────────────────────────────────────
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9           # PostgreSQL driver

# ── Task Queue ────────────────────────────────────────────
celery[redis]>=5.3.6
redis>=5.0.8

# ── Object Storage ────────────────────────────────────────
minio>=7.2.9

# ── Search ────────────────────────────────────────────────
elasticsearch>=8.15.0

# ── Document Extraction ───────────────────────────────────
pdfplumber>=0.11.4
pymupdf>=1.24.10                 # fitz — image detection
python-docx>=1.1.2               # DOCX extraction

# ── ML / Embeddings / Reranker ────────────────────────────
# NOTE: Install PyTorch FIRST before installing sentence-transformers
# Run this command separately (CPU-only, recommended for dev):
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
#
# Then install the rest:
sentence-transformers>=3.1.1     # wraps bge-large + bge-reranker
transformers>=4.45.0             # tokenizer for chunking
huggingface-hub>=0.25.0          # model downloads

# ── LLM ───────────────────────────────────────────────────
openai>=1.52.0                   # OpenAI-compatible client (works with OpenRouter)
httpx>=0.27.2                    # async HTTP

# ── Config & Validation ───────────────────────────────────
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1

# ── Utilities ─────────────────────────────────────────────
python-jose>=3.3.0               # JWT (reserved for future auth)
